fix-possible-integer-overflow-tnefpropertyreader

+++ MimeKit/Tnef/TnefPropertyReader.cs
@@ -302,7 +302,7 @@ namespace MimeKit.Tnef {
 				case TnefPropertyType.Binary:
 				case TnefPropertyType.Object:
 					int n = ReadInt32 ();
-					if (n >= 0 && n + 4 < length)
+					if (n >= 0 && (long)n + 4 < length)
 						length = n + 4;
 					break;
 				}
